Refactor out part of the sandbox code so it can be used for Jvm entrypoint.
