+++ korge-sandbox/build.gradle
@@ -2,6 +2,10 @@ import korlibs.korge.gradle.targets.jvm.KorgeJavaExec
 
 description = "Multiplatform Game Engine written in Kotlin"
 
+korge {
+    entrypoint("JvmMain", "JvmMain")
+}
+
 dependencies {
 	add("commonMainApi", project(":korge"))
 }
