+++ korge-sandbox/src/commonMain/kotlin/Main.kt
@@ -135,8 +135,7 @@ suspend fun main() = Korge(
         //Demo(::MainColorTransformFilter),
         //Demo(::MainMasks),
         //Demo(::MainShape2dScene),
-        //Demo(::MainUIStacks),
-        Demo(::MainCache),
+        Demo(::MainUIStacks),
         //Demo(::MainSprites10k),
         //Demo(::MainStressMatrixMultiplication),
         //Demo(::MainSDF),
@@ -242,45 +241,4 @@ suspend fun main() = Korge(
     //text("%#HELLO WORLD!", textSize = 65f, font = helvetica).xy(100, 100)
 }
 
-class Demo(val sceneBuilder: () -> Scene, val name: String = sceneBuilder()::class.portableSimpleName.removePrefix("Main")) {
-    override fun toString(): String = name
-}
 
-suspend fun Stage.demoSelector(default: Demo, all: List<Demo>) {
-    val container = sceneContainer(size = Size(width, height - 48f)) { }.xy(0, 48)
-    val containerFocus = container.makeFocusable()
-
-    lateinit var comboBox: UIComboBox<Demo>
-
-    suspend fun setDemo(demo: Demo?) {
-        //container.removeChildren()
-        if (demo != null) {
-            comboBox.selectedItem = demo
-            views.clearColor = DEFAULT_KORGE_BG_COLOR
-            container.changeTo {
-                containerFocus.focus()
-                demo.sceneBuilder().also { it.init(this) }
-            }
-        }
-    }
-
-    uiHorizontalStack(padding = 8f) {
-        alignLeftToLeftOf(this@demoSelector.stage, padding = 8.0).alignTopToTopOf(this@demoSelector.stage, padding = 8.0)
-        comboBox = uiComboBox<Demo>(size = UI_DEFAULT_SIZE.copy(width = 200f), items = (listOf(default) + all).distinctBy { it.name }.sortedBy { it.name }) {
-            this.viewportHeight = 600
-            this.onSelectionUpdate.add {
-                //println(it)
-                launchImmediately { setDemo(it.selectedItem!!) }
-            }
-        }
-        uiCheckBox(size = UI_DEFAULT_SIZE.copy(width = 200f), text = "forceRenderEveryFrame", checked = views.forceRenderEveryFrame) {
-            onChange { views.forceRenderEveryFrame = it.checked }
-        }
-        uiCheckBox(size = UI_DEFAULT_SIZE.copy(width = 150f), text = "toggleDebug", checked = views.debugViews) {
-            onChange { views.debugViews = it.checked }
-        }
-    }
-    comboBox.selectedItem = default
-    comboBox.focusNoOpen()
-    //setDemo(default)
-}
